consule service: pass config node attributes to consul flags
data_dir is defined in the node config attributes. The default corresponds to
the default in consul_service, and consul_service creates the directory. If the
default is changed, consul_service needs to know of this non-default value.

config_dir is defined in attributes but not passed to consul_service to
override the default.

consul_service needs to restart if an existing cert changes - if CACert or Key
changes, the Cert will change.